@@ -1,9 +1,8 @@
-﻿namespace RenderTableCreator
+﻿namespace RenderTableCreator;
+
+/// <summary>
+/// Interaction logic for App.xaml
+/// </summary>
+public partial class App
 {
-    /// <summary>
-    /// Interaction logic for App.xaml
-    /// </summary>
-    public partial class App
-    {
-    }
 }
